Remove remaining non-SDL usage of SDL fixed width int types.

+++ arch/nds/platform.c
@@ -40,15 +40,15 @@ extern void nds_update_input(void);
 #define timers2ms(tlow,thigh) ((tlow) | ((thigh)<<16)) >> 7
 #define timers2ticks(tlow,thigh) ((tlow) | ((thigh)<<16))
 
-void delay(Uint32 ms)
+void delay(uint32_t ms)
 {
-  Uint32 now;
+  uint32_t now;
   now = timers2ms(TIMER0_DATA, TIMER1_DATA);
-  while((Uint32)timers2ms(TIMER0_DATA, TIMER1_DATA) < now + ms)
+  while((uint32_t)timers2ms(TIMER0_DATA, TIMER1_DATA) < now + ms)
     ;
 }
 
-Uint32 get_ticks(void)
+uint64_t get_ticks(void)
 {
   return timers2ms(TIMER0_DATA, TIMER1_DATA);
 }
